pyproject.toml
[tool.poetry]
name = "personal-website-streamlit"
version = "0.1.0"
description = ""
authors = ["pythoninthegrass <4097471+pythoninthegrass@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "3.10.4"
streamlit-lottie = "0.0.2"
streamlit = "1.3.0"
requests = "2.24.0"
Pillow = "8.4.0"
protobuf = "~=3.19.0"
watchdog = "^2.1.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"